requirements.txt
asgiref==3.5.2
Django==4.1.1
djangorestframework==3.13.1
flake8==5.0.4
mccabe==0.7.0
pycodestyle==2.9.1
pyflakes==2.5.0
pytz==2022.2.1
sqlparse==0.4.2
tzdata==2022.2